refactor: GameRoom과 BoardState 책임 분리
